@@ -1,7 +1,9 @@
-const { Kafka, logLevel: kafkaLogLevels } = require('kafkajs')
+import { Kafka, logLevel as kafkaLogLevels } from 'kafkajs'
 
-const globalLogger = require('../logger')
-const { KAFKA_BROKERS, KAFKA_PAYLOAD_TOPIC, KAFKA_LOG_LEVEL } = require('../env')
+import globalLogger from '../logger.js'
+import env from '../env.js'
+
+const { KAFKA_BROKERS, KAFKA_PAYLOAD_TOPIC, KAFKA_LOG_LEVEL } = env
 
 const setupForwardToKafka = async () => {
   const logger = globalLogger.child({ module: 'Kafka' })
@@ -35,4 +37,4 @@ const setupForwardToKafka = async () => {
   }
 }
 
-module.exports = setupForwardToKafka
+export default setupForwardToKafka
