@@ -1,5 +1,7 @@
-const uuid = require('uuid')
-const { WASP_INGEST_NAME } = require('../env')
+import { v4 } from 'uuid'
+import env from '../env.js'
+
+const { WASP_INGEST_NAME } = env
 
 const setupParser = (next) => {
   return (topic, payload) => {
@@ -8,7 +10,7 @@ const setupParser = (next) => {
       throw new Error(`Invalid ID format in topic: ${topic}`)
     }
     const hardwareSerial = result[1]
-    const payloadId = uuid.v4()
+    const payloadId = v4()
 
     next({
       ingestId: hardwareSerial,
@@ -26,4 +28,4 @@ const setupParser = (next) => {
   }
 }
 
-module.exports = setupParser
+export default setupParser
